Added code that does post render
Added interface which does a post render onto the bitmap before it's
presented to the screen. This will be useful for printing FSP stats or
doing some non-graphical work in the graphics thread.

It's also possible to split the work up between threads so that work is
processed on a different thread, while graphics runs as fast as possible
on the graphics thread. It will need more code and more work, but it
will be nice to have.